removed replacement for AssHString
Procedure CgAssOut.AssHString was specified as emitting the entire character array to the assembly file (from index 0 to HIGH). However, this turned out to be incorrect. AssHString is semantically equivalent to AssString. The only difference is premature optimisation within AssHString (iterating over chunks of four characters instead of every character). The code is so obviously confusing that even the authors got it wrong. Yet, the procedure is predominantly used to write short strings of under 10 chars, and at most 20 something chars. There is no need for such a procedure at all. Consequently its replacement EmitArrayOfChar is also obsolete.